Add UI tests for user registration, login, exercise management, goal progress, and share page functionality
- Implemented `TestBrowseAddExerciseAndVisualize` to test browsing, adding exercises, and visualizing data.
- Created `TestDashboardGoalProgress` to verify goal progress updates after adding exercises.
- Developed `TestRegisterErrors` to check for error messages during user registration with duplicate emails and missing fields.
- Added `SystemTests` for comprehensive testing of user registration, login, and logout processes.
- Introduced `TestSharePageUI` to validate the UI elements and modals on the share page.
